LayoutTests/svg/dom/SVGAnimatedInteger-expected.txt

This test checks the SVGAnimatedInteger API - utilizing the orderX property of feConvolveMatrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial filterResX value
PASS convolveMatrix.orderX.toString() is "[object SVGAnimatedInteger]"
PASS typeof(convolveMatrix.orderX.baseVal) is "number"
PASS convolveMatrix.orderX.baseVal is 0

Check that integers are static, caching value in a local variable and modifying it, should have no effect
PASS numRef is 100
PASS convolveMatrix.orderX.baseVal is 0

Check assigning various valid and invalid values
PASS convolveMatrix.orderX.baseVal = -1 is -1
PASS convolveMatrix.orderX.baseVal = 300 is 300
PASS convolveMatrix.orderX.baseVal = 'aString' is 'aString'
PASS convolveMatrix.orderX.baseVal is 0
PASS convolveMatrix.orderX.baseVal = convolveMatrix is convolveMatrix
PASS convolveMatrix.orderX.baseVal is 0
PASS convolveMatrix.orderX.baseVal = 300 is 300

Check that the filterResX value remained 300
PASS convolveMatrix.orderX.baseVal is 300
PASS successfullyParsed is true

TEST COMPLETE
